refactor: improve memory management and precision in download handling
- Fix division by zero in percentage calculation using double precision
- Prevent infinite loop in CompleteDownloadDialog closeEvent by removing redundant close()
- Parameterize completed status ID in database queries for maintainability
- Add null checks and Qt::WA_DeleteOnClose attribute in ShowCompleteDialog
- Implement proper cleanup in Download and MainWindow destructors

+++ src/ProcessDatabaseOutput.cpp
@@ -95,8 +95,9 @@ void ProcessDatabaseOutput::ProcessPrepareLoadedInformationForMainTableView(cons
 	}
 	else
 	{
-		float Present = (long double)DownloadedSize / DownloadSize;
-		Status = QString::number(Present*100, 'f', 2)+"%";
+		// Use double precision for accurate percentage calculation
+		double Present = static_cast<double>(DownloadedSize) / static_cast<double>(DownloadSize);
+		Status = QString::number(Present * 100.0, 'f', 2) + "%";
 	}
